+++ lib/src/utils/responsive.dart
@@ -5,33 +5,33 @@ import 'package:shadcn_ui/src/theme/theme.dart';
 
 class ShadBreakpoints {
   ShadBreakpoints({
-    double tiny = 0,
-    double small = 640,
-    double medium = 768,
-    double large = 1024,
-    double extraLarge = 1280,
-    double extraExtraLarge = 1536,
-  })  : tiny = ShadBreakpointTiny(tiny),
-        small = ShadBreakpointSmall(small),
-        medium = ShadBreakpointMedium(medium),
-        large = ShadBreakpointLarge(large),
-        extraLarge = ShadBreakpointExtraLarge(extraLarge),
-        extraExtraLarge = ShadBreakpointExtraExtraLarge(extraExtraLarge);
+    double tn = 0,
+    double sm = 640,
+    double md = 768,
+    double lg = 1024,
+    double xl = 1280,
+    double xxl = 1536,
+  })  : tn = ShadBreakpointTN(tn),
+        sm = ShadBreakpointSM(sm),
+        md = ShadBreakpointMD(md),
+        lg = ShadBreakpointLG(lg),
+        xl = ShadBreakpointXL(xl),
+        xxl = ShadBreakpointXXL(xxl);
 
-  final ShadBreakpoint tiny;
-  final ShadBreakpoint small;
-  final ShadBreakpoint medium;
-  final ShadBreakpoint large;
-  final ShadBreakpoint extraLarge;
-  final ShadBreakpoint extraExtraLarge;
+  final ShadBreakpoint tn;
+  final ShadBreakpoint sm;
+  final ShadBreakpoint md;
+  final ShadBreakpoint lg;
+  final ShadBreakpoint xl;
+  final ShadBreakpoint xxl;
 
   ShadBreakpoint fromWidth(double width) {
-    if (width < small.value) return tiny;
-    if (width < medium.value) return small;
-    if (width < large.value) return medium;
-    if (width < extraLarge.value) return large;
-    if (width < extraExtraLarge.value) return extraLarge;
-    return extraExtraLarge;
+    if (width < sm.value) return tn;
+    if (width < md.value) return sm;
+    if (width < lg.value) return md;
+    if (width < xl.value) return lg;
+    if (width < xxl.value) return xl;
+    return xxl;
   }
 
   static ShadBreakpoints lerp(
@@ -40,13 +40,12 @@ class ShadBreakpoints {
     double t,
   ) {
     return ShadBreakpoints(
-      tiny: ShadBreakpoint.lerp(a!.tiny, b!.tiny, t),
-      small: ShadBreakpoint.lerp(a.small, b.small, t),
-      medium: ShadBreakpoint.lerp(a.medium, b.medium, t),
-      large: ShadBreakpoint.lerp(a.large, b.large, t),
-      extraLarge: ShadBreakpoint.lerp(a.extraLarge, b.extraLarge, t),
-      extraExtraLarge:
-          ShadBreakpoint.lerp(a.extraExtraLarge, b.extraExtraLarge, t),
+      tn: ShadBreakpoint.lerp(a!.tn, b!.tn, t),
+      sm: ShadBreakpoint.lerp(a.sm, b.sm, t),
+      md: ShadBreakpoint.lerp(a.md, b.md, t),
+      lg: ShadBreakpoint.lerp(a.lg, b.lg, t),
+      xl: ShadBreakpoint.lerp(a.xl, b.xl, t),
+      xxl: ShadBreakpoint.lerp(a.xxl, b.xxl, t),
     );
   }
 }
@@ -77,28 +76,28 @@ sealed class ShadBreakpoint {
   }
 }
 
-class ShadBreakpointTiny extends ShadBreakpoint {
-  const ShadBreakpointTiny(super.value);
+class ShadBreakpointTN extends ShadBreakpoint {
+  const ShadBreakpointTN(super.value);
 }
 
-class ShadBreakpointSmall extends ShadBreakpoint {
-  const ShadBreakpointSmall(super.value);
+class ShadBreakpointSM extends ShadBreakpoint {
+  const ShadBreakpointSM(super.value);
 }
 
-class ShadBreakpointMedium extends ShadBreakpoint {
-  const ShadBreakpointMedium(super.value);
+class ShadBreakpointMD extends ShadBreakpoint {
+  const ShadBreakpointMD(super.value);
 }
 
-class ShadBreakpointLarge extends ShadBreakpoint {
-  const ShadBreakpointLarge(super.value);
+class ShadBreakpointLG extends ShadBreakpoint {
+  const ShadBreakpointLG(super.value);
 }
 
-class ShadBreakpointExtraLarge extends ShadBreakpoint {
-  const ShadBreakpointExtraLarge(super.value);
+class ShadBreakpointXL extends ShadBreakpoint {
+  const ShadBreakpointXL(super.value);
 }
 
-class ShadBreakpointExtraExtraLarge extends ShadBreakpoint {
-  const ShadBreakpointExtraExtraLarge(super.value);
+class ShadBreakpointXXL extends ShadBreakpoint {
+  const ShadBreakpointXXL(super.value);
 }
 
 typedef ResponsiveWidgetBuilder = Widget Function(